+++ src/Psalm/Exception/CircularReferenceException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Psalm\Exception;
+
+class CircularReferenceException extends CodeException
+{
+}
